Fixed warnings detected by clang

+++ src/tree/FenwickTree.h
@@ -34,7 +34,7 @@ class FenwickTree {
    */
   explicit FenwickTree(const std::vector<T>& nums)
       : nums_(nums.size(), 0), bits_(nums_.size() + 1, 0) {
-    for (int i = 0; i < nums_.size(); ++i) {
+    for (size_t i = 0; i < nums_.size(); ++i) {
       update(i, nums[i]);
     }
   }
